% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate the index data set}
\usage{
aggregate_data(dat)
}
\arguments{
\item{dat}{The index data set as output from \code{get_index_data}.}
}
\value{
A \code{tibble} of aggregated data from \code{dat}.
}
\description{
For modelling purposes, the data should be aggregated according to
\code{randomisation_outcome} and \code{current_eligible_vaccination}
which will be covariates in the model.
}
