% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{brar}
\alias{brar}
\title{Bayesian response adaptive randomisation}
\usage{
brar(pbest, sampsize, variance, inactive, fix_ctrl = 1/13)
}
\arguments{
\item{pbest}{Probability arm is best}

\item{sampsize}{Current sample size allocated to arm}

\item{variance}{Current posterior variance}

\item{inactive}{Flag for if an arm should receive zero allocations}

\item{fix_ctrl}{Fix allocation to control by this amount}
}
\value{
A numeric vector giving the BRAR allocation probabilities
}
\description{
Bayesian response adaptive randomisation
}
