% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{brar1}
\alias{brar1}
\title{Bayesian adaptive randomisation method 1}
\usage{
brar1(p, no_alloc_thres = 0.01, fix_ctrl = NULL)
}
\arguments{
\item{p}{The probability each arm is best}

\item{no_alloc_thres}{The threshold for setting allocations to zero}

\item{fix_ctrl}{The fixed allocation amount to control}
}
\value{
A vector of allocation probabilities for each arm
}
\description{
Bayesian adaptive randomisation method 1
}
