% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{brar3}
\alias{brar3}
\title{Bayesian adaptive randomisation method 3}
\usage{
brar3(p, n, V, no_alloc_thres = 0.01, fix_ctrl = NULL, min_alloc = 0)
}
\arguments{
\item{p}{The probability each arm is best}

\item{n}{The current sample size in each arm}

\item{V}{The current posterior variance in each arm}

\item{no_alloc_thres}{The threshold for setting allocations to zero}

\item{fix_ctrl}{The fixed allocation amount to control}

\item{min_alloc}{The minimum allocation to an arm}
}
\value{
A vector of allocation probabilities for each arm
}
\description{
Bayesian adaptive randomisation method 3
}
