% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{col_comp}
\alias{col_comp}
\title{Calcualte probability that each column is superior to a reference column.}
\usage{
col_comp(mat, col, eps = 0)
}
\arguments{
\item{mat}{Matrix of parameter draws}

\item{col}{The reference column integer}

\item{eps}{The reference value for superiority, default is \code{0}}
}
\value{
A vector of probabilities
}
\description{
Calcualte probability that each column is superior to a reference column.
}
