% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{col_sup}
\alias{col_sup}
\title{Calculate the probability that a column is superior to all other columns}
\usage{
col_sup(mat, col, eps = 0)
}
\arguments{
\item{mat}{Matrix of parameter draws}

\item{col}{The reference column}

\item{eps}{The value for superiority}
}
\value{
A numeric value
}
\description{
Calculate the probability that a column is superior to all other columns
}
