% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{col_sup_all}
\alias{col_sup_all}
\title{Calculate the probability that each column is superior to all other columns}
\usage{
col_sup_all(mat, eps = 0)
}
\arguments{
\item{mat}{Matrix of parameter draws}

\item{eps}{The value for superiority}
}
\value{
A vector of probabilities
}
\description{
Calculate the probability that each column is superior to all other columns
}
