% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{constrain_design}
\alias{constrain_design}
\title{Eigen-decompose a constraining prior to create identifiable parameters}
\usage{
constrain_design(X, S)
}
\arguments{
\item{X}{The design matrix}

\item{S}{The constraint matrix}
}
\value{
The constrained design matrix
}
\description{
Eigen-decompose a constraining prior to create identifiable parameters
}
