% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{distribute_alloc}
\alias{distribute_alloc}
\title{Re-distribute allocations}
\usage{
distribute_alloc(w, zero_ind)
}
\arguments{
\item{w}{The current allocation weights}

\item{zero_ind}{Indices for arms which will receive zero allocation}
}
\value{
An updated vector of allocation weights
}
\description{
When some allocations a set to zero for being below some threshold,
re-distribute their mass amongst the remaining arms
}
