% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate model parameters}
\usage{
estimate_model(agg_dat, seed, ...)
}
\arguments{
\item{agg_dat}{Aggregated data from \code{aggregate_data}}

\item{seed}{PRNG seed for Stan fit}

\item{...}{Other arguments to brms/sampling.}
}
\value{
A Stan model object
}
\description{
Estimate model parameters
}
