% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_allocation_sequence}
\alias{generate_allocation_sequence}
\title{Generate an allocation sequence}
\usage{
generate_allocation_sequence(num_alloc, alloc_prob, seed)
}
\arguments{
\item{num_alloc}{Number of allocations to generate, i.e. length of sequence}

\item{alloc_prob}{Allocation probability to each arm}

\item{seed}{Seed for PRNG}
}
\value{
A numeric vector sequence of arm allocations
}
\description{
Generate an allocation sequence
}
