% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_params.R
\name{get_design_matrix}
\alias{get_design_matrix}
\title{Return a matrix giving the model design}
\usage{
get_design_matrix()
}
\value{
A design matrix
}
\description{
Return a matrix giving the model design
}
