% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_index_data}
\alias{get_index_data}
\title{Return index records (i.e. first record) for \code{parent_id}}
\usage{
get_index_data(dat)
}
\arguments{
\item{dat}{The processed raw data; a \code{tibble}.}
}
\value{
A \code{tibble} subset of \code{dat}
  with one row per \code{parent_id} corresponding to the earliest
  due vaccination.
}
\description{
Primary analysis will be based on only index records for \code{parent_id}.
}
