% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_params.R
\name{get_interim_schedule}
\alias{get_interim_schedule}
\title{Return a \code{data.frame} giving the interim schedule and seeds.}
\usage{
get_interim_schedule()
}
\value{
A \code{data.frame} giving the interim schedule and seeds.
}
\description{
Return a \code{data.frame} giving the interim schedule and seeds.
}
