% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_params.R
\name{get_interim_seeds}
\alias{get_interim_seeds}
\title{Return a numeric vector giving the seeds for an interim.}
\usage{
get_interim_seeds(interim)
}
\arguments{
\item{interim}{The interim number for which seeds are required.}
}
\value{
A numeric vector giving the allocation seed then the Stan seed.
}
\description{
Return a numeric vector giving the seeds for an interim.
}
