% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_params.R
\name{get_intervention_map}
\alias{get_intervention_map}
\title{Return intervention mapping data.frame}
\usage{
get_intervention_map()
}
\value{
A \code{data.frame} giving the intervention to arm mapping.
}
\description{
Return intervention mapping data.frame
}
