% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_posterior_quantities}
\alias{get_posterior_quantities}
\title{Calculate probability best}
\usage{
get_posterior_quantities(mod)
}
\arguments{
\item{mod}{Stan model}
}
\value{
Numeric vector giving probability each arm is best
}
\description{
Calculate probability best
}
