% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_params.R
\name{get_trial_params}
\alias{get_trial_params}
\title{Return list of trial parameters}
\usage{
get_trial_params()
}
\value{
A \code{list} giving trial parameters.
}
\description{
Return list of trial parameters
}
