% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init_interim_log}
\alias{init_interim_log}
\title{Intialise the interim log}
\usage{
init_interim_log(file)
}
\arguments{
\item{file}{The path for the interim log file}
}
\description{
Intialises the interim log based on \code{trial_params}.
Should only be called on first build.
}
