% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_model_data}
\alias{make_model_data}
\title{Make data list for input into Stan model}
\usage{
make_model_data(agg_dat)
}
\arguments{
\item{agg_dat}{Aggregated data as outcome form aggregate_data.}
}
\value{
A list of data for input int \code{rstan::sampling}.
}
\description{
Make data list for input into Stan model
}
