% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{pairwise_comp}
\alias{pairwise_comp}
\title{Make a pairwise comparison between parameter draws}
\usage{
pairwise_comp(mat, eps = 0)
}
\arguments{
\item{mat}{Matrix of parameter draws}

\item{eps}{Reference value, default is \code{0}.}
}
\value{
Vector of probabilities that pairwise difference is greater than \code{eps}.
}
\description{
Make a pairwise comparison between parameter draws
}
