% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{pairwise_diff}
\alias{pairwise_diff}
\title{Calculate matrix of pairwise differences of MC draws}
\usage{
pairwise_diff(mat, trans = I)
}
\arguments{
\item{mat}{The matrix of draws}

\item{trans}{A transformation function, default is identity \code{I()}.}
}
\value{
A matrix of pairwise differences
}
\description{
Calculate matrix of pairwise differences of MC draws
}
