% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{prob_all_equivalent}
\alias{prob_all_equivalent}
\title{Probability all columns are equivalent - spherical}
\usage{
prob_all_equivalent(mat, eps)
}
\arguments{
\item{mat}{Matrix of MC draws}

\item{eps}{Equivalence bound}
}
\value{
The probability of spherical equivalence
}
\description{
Spherical test that all columns are within bound delta of global mean
}
