% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{prob_all_noninferior}
\alias{prob_all_noninferior}
\title{Probability all columns are non-inferior to best column}
\usage{
prob_all_noninferior(mat, colvec_best, eps)
}
\arguments{
\item{mat}{Matrix of MC draws}

\item{colvec_best}{A vector of draws from the best}

\item{eps}{The reference value for non-inferiority}
}
\value{
A real number giving the probability that all arms
  are non-inferior to colvec_best.
}
\description{
Probability all columns are non-inferior to best column
}
