% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{prob_best}
\alias{prob_best}
\title{Probability a column is maximum}
\usage{
prob_best(mat)
}
\arguments{
\item{mat}{Matrix of MC draws}
}
\value{
Numeric vector giving probability each column is the maximum
}
\description{
Probability a column is maximum
}
