% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{prob_best_lin}
\alias{prob_best_lin}
\title{Probability best of linear transformation}
\usage{
prob_best_lin(mat, A)
}
\arguments{
\item{mat}{Matrix of MC draws}

\item{A}{Transformation matrix}
}
\value{
Numeric vector giving probability best each transformed column is maximum
}
\description{
For a supplied linear transform, A, of MC draws, determine
the probability best.
}
