% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{prob_equi}
\alias{prob_equi}
\title{Probability all columns are equivalent - spherical}
\usage{
prob_equi(mat, delta, ...)
}
\arguments{
\item{mat}{Matrix of MC draws}

\item{delta}{Equivalence bound}
}
\value{
The probability of spherical equivalence
}
\description{
Spherical test that all columns are within bound delta of global mean
}
