% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_functions.R
\name{prob_rank}
\alias{prob_rank}
\title{Probability for each rank}
\usage{
prob_rank(mat)
}
\arguments{
\item{mat}{Matrix of MC draws}
}
\value{
Numeric square matrix giving probability each arm is ranked in each place
}
\description{
Rank each row and return the probability that each arm is ranked at each value
}
