% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Process raw data}
\usage{
process_raw_data(raw_dat)
}
\arguments{
\item{raw_dat}{The raw data.}
}
\value{
A \code{data.table}.
}
\description{
Needs to check raw data for any coding issues etc.
}
