% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_interim_log}
\alias{read_interim_log}
\title{Read interim log file.}
\usage{
read_interim_log(file)
}
\arguments{
\item{file}{Path to interim log.}
}
\value{
A \code{data.table} of interim log.
}
\description{
Read interim log file.
}
