% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read raw data}
\usage{
read_raw_data(file)
}
\arguments{
\item{file}{The path to the database data file.}
}
\value{
A \code{tibble}.
}
\description{
Read raw data
}
