% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{update_alloc}
\alias{update_alloc}
\title{Update allocation probabilities if some must be set to zero}
\usage{
update_alloc(p, zero_ind)
}
\arguments{
\item{p}{The allocation probabilities}

\item{zero_ind}{The indices of values to be set to zero}
}
\value{
Updated allocation probabilities
}
\description{
Update allocation probabilities if some must be set to zero
}
