% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_interim_log}
\alias{update_interim_log}
\title{Update the interim log file}
\usage{
update_interim_log(file, date, interim, n_analysed, prob_alloc, inactive)
}
\arguments{
\item{file}{The path to the interim log file}

\item{date}{The date of the interim analysis}

\item{interim}{The interim sequence number}

\item{n_analysed}{The sample size analysed}

\item{prob_alloc}{The allocation probability vector}

\item{is_alloc}{Flag indicating arms which will receive allocations up to the next interim.}
}
\description{
The purpose of the interim log file is to track the current
interim, allocation probabilities, and allocation indicators,
as well as record when interims were undertaken and with
what sample size.
}
