% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_allocation_sequence}
\alias{write_allocation_sequence}
\title{Write allocation sequence to file}
\usage{
write_allocation_sequence(file, alloc_seq, interim)
}
\arguments{
\item{file}{Filepath for allocation sequence file}

\item{alloc_seq}{The allocation sequence}

\item{interim}{Integer giving interim from which allocation was generated}
}
\description{
Write allocation sequence to file
}
