% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{brar}
\alias{brar}
\title{Bayesian response adaptive randomisation}
\usage{
brar(pbest, sampsize, variance, no_alloc_thres, fix_ctrl = NULL)
}
\arguments{
\item{pbest}{Probability arm is best}

\item{sampsize}{Current sample size allocated to arm}

\item{variance}{Current posterior variance}

\item{no_alloc_thres}{Threshold for setting allocation to zero}

\item{fix_ctrl}{Fix allocation to control by this amount}
}
\value{
A numeric vector giving the BRAR allocation probabilities
}
\description{
Bayesian response adaptive randomisation
}
