% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_data.R
\name{generate_trial_data}
\alias{generate_trial_data}
\title{Create simulated data for testing purposes}
\usage{
generate_trial_data(n, seed)
}
\arguments{
\item{n}{Number of rows to simulate.}

\item{seed}{Seed for PRNG.}
}
\value{
A \code{data.frame} of simulated data.
}
\description{
Ideally this will cover all error/coding issues we
might expect to see so we can check our handling
}
