% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_posterior_quantities}
\alias{get_posterior_quantities}
\title{Calculate probability best}
\usage{
get_posterior_quantities(draws, sampsize)
}
\arguments{
\item{draws}{MC model draws for arm means}

\item{sampsize}{Sample size for each arm}
}
\value{
A \code{tibble} giving posterior quantities of interest for each arm.
}
\description{
Calculate probability best
}
