% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Process raw data}
\usage{
process_raw_data(raw_dat, ref_date)
}
\arguments{
\item{raw_dat}{The raw data.}

\item{ref_date}{The reference date from which to calculate how much time has passed since the vaccine due date.
This should probably be the date on which the function was called, i.e. \code{Sys.Date()}.}
}
\value{
A \code{tibble}.
}
\description{
Needs to check raw data for any coding issues etc.
}
